pyproject.toml
[project]
name = "azure-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "azure-servicebus>=7.14.2",
    "python-dotenv>=1.1.1",
]
